==> main/00_pandas_cheat_sheet.ipynb <==
## Pandas Cheat Sheet for Beginners

### Importing pandas

import pandas as pd

### Creating DataFrames

# From a dictionary
data = {'Column1': [1, 2, 3], 'Column2': [4, 5, 6]}
dfd = pd.DataFrame(data)

# From a list of lists
data = [[1, 2, 3], [4, 5, 6]]
dfl = pd.DataFrame(data, columns=['Column1', 'Column2', 'Column3'])

dfd # DataFrame from a dictionary

dfl # DataFrame from a list of lists

### Reading and Writing Data

# Reading data from a CSV file
# df = pd.read_csv('file.csv')

# Writing data to a CSV file
# df.to_csv('file.csv', index=False)

# Reading data from an Excel file
# df = pd.read_excel('file.xlsx')

# Writing data to an Excel file
# df.to_excel('file.xlsx', index=False)

### Viewing Data

# Creating a simple DataFrame
data = {
    'Player': ['LeBron James', 'Stephen Curry', 'Kevin Durant', 'Giannis Antetokounmpo'],
    'Points': [27, 24, 29, 28],
    'Rebounds': [7, 5, 8, 11],
    'Assists': [8, 6, 5, 6]
}

df = pd.DataFrame(data)

# Display the first few rows
df.head(1)

# Display the last few rows
df.tail(1)

# Display basic information about the DataFrame
df.info()

# Display summary statistics
df.describe()

### Selecting Data

# Selecting a single column
df['Points']

# Selecting multiple columns
df[['Rebounds', 'Assists']]


# Selecting rows by index
df.iloc[0]  # First row

# Selecting rows by index
df.iloc[0:2]  # First two rows

# Selecting rows by condition
df[df['Rebounds'] > 7]

### Modifying Data

# Adding a new column
df['Total'] = df['Points'] + df['Rebounds'] + df['Assists']
df[df['Player']=='Stephen Curry']


# Renaming columns
df.rename(columns={'Total': 'Pts+Rbs+Ats'}, inplace=True)
df.iloc[:1]

# Dropping a column
df.drop('Pts+Rbs+Ats', axis=1, inplace=True)
df.iloc[2:]

### Handling Missing Data

# Adding a new player row
new_player_df = pd.DataFrame([{'Player': 'Bronny James', 'Points': 0, 'Rebounds': 0, 'Assists': None}])
df = pd.concat([df, new_player_df], ignore_index=True)
df

# Checking for missing values
df.isnull().sum()

df[df['Assists'].notna()]

# Show rows where the Assists cell has NaN/None value
df[df['Assists'].isna()]

# Dropping rows with missing values
df.dropna(inplace=True)

# Filling missing values
# df.fillna(0, inplace=True)
df

### Grouping and Aggregating Data

# Grouping data by a column and calculating the mean
# df.groupby('Column1').mean()

# Aggregating data with multiple functions
# df.groupby('Column1').agg({'Column2': ['mean', 'sum']})

### Misc

round(df['Points'].mean(), 2)

df['Points'].sum()

df.columns

df.shape

df['Player'] = df['Player'].replace('Giannis Antetokounmpo', 'Greek Freak')
df[df['Player'] == 'Greek Freak']

df.Player.to_list()

---
