==> main/01_extract_data.ipynb <==
Imports

import pandas as pd
import requests
pd.set_option('display.max_columns', None) # displays all columns in a wide DataFrame
import time
import numpy as np

Test Data Extraction

test_url = 'https://stats.nba.com/stats/leagueLeaders?LeagueID=00&PerMode=Totals&Scope=S&Season=2012-13&SeasonType=Regular%20Season&StatCategory=PTS'
r = requests.get(url=test_url).json()
table_headers = r['resultSet']['headers']
pd.DataFrame(r['resultSet']['rowSet'], columns=table_headers).head()

temp_df1 = pd.DataFrame(r['resultSet']['rowSet'], columns=table_headers)
temp_df2 = pd.DataFrame({
    'Year': ['2023-24' for i in range(len(temp_df1))],
    'Season_type': ['Regular%20Season' for i in range(len(temp_df1))],
})
temp_df3 = pd.concat([temp_df2, temp_df1], axis=1)

temp_df3.head()

del temp_df1, temp_df2, temp_df3

Table Headers Setup

df_cols = ['Year', 'Season_type'] + table_headers

pd.DataFrame(columns=df_cols)

Extract Data & Export to Excel

df = pd.DataFrame(columns=df_cols)
season_types = ['Regular%20Season', 'Playoffs']
years = ['2012-13', '2013-14', '2014-15', '2015-16', '2016-17', '2017-18', '2018-19', '2019-20', '2020-21', '2021-22', '2022-23', '2023-24']

start_data_extraction = time.time()

for y in years:
    for s in season_types:
        api_url = f'https://stats.nba.com/stats/leagueLeaders?LeagueID=00&PerMode=Totals&Scope=S&Season={y}&SeasonType={s}&StatCategory=PTS'
        r = requests.get(url=api_url).json()
        temp_df1 = pd.DataFrame(r['resultSet']['rowSet'], columns=table_headers)
        temp_df2 = pd.DataFrame({
            'Year': [y for i in range(len(temp_df1))],
            'Season_type': [s for i in range(len(temp_df1))],
        })
        temp_df3 = pd.concat([temp_df2, temp_df1], axis=1)
        df = pd.concat([df, temp_df3], axis=0)
        print(f'Finished collecting data for the {y} {s}.')
        lag = np.random.uniform(low=5.0, high=25.0) # this lag is not mandatory
        print(f'... waiting {round(lag, 1)} seconds until next attempt.')
        time.sleep(lag) # this lag is not mandatory

print(f'Process completed! Total run time: {round((time.time() - start_data_extraction)/60,2)} minutes.')

# df.to_excel('nba_player_data.xlsx', index=False) # UNCOMMENT if you want TO CREATE a NEW nba player data SPREADSHEET!!!

---
