==> main/05_gsw_analyze_player_data.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)


df = pd.read_excel('../data/gsw_box_score_player_stats.xlsx')

### Stephen Curry Stats

steph_df = df[(df['NAME'] == 'Stephen Curry') & (df['MIN'].notna())]
steph_df.NAME.nunique()

steph_df.tail(3)

def get_best_or_worst_player_stats(stat: str, lowest: bool, other_columns: list, n: int=5):
    filtered_games = steph_df.sort_values(by=stat, ascending=lowest).head(n)
    print(f"\nSteph Curry's average {stat} per game this season: {round(steph_df[stat].mean(),2)}")
    print(f"Steph Curry's average {stat} based on this {n}-game sample: {filtered_games[stat].mean()}")
    print(filtered_games[['TEAM', 'OPP', 'STATUS', 'DATE', 'MIN', stat] + other_columns].sort_values(stat, ascending=lowest))
    return filtered_games

steph_lowest_pts_games = get_best_or_worst_player_stats(stat='PTS', lowest=True, other_columns=['FG%', '3P%', 'FT%'], n=5)
steph_highest_pts_games = get_best_or_worst_player_stats(stat='PTS', lowest=False, other_columns=['FG%', '3P%', 'FT%'], n=5)

steph_lowest_3pt_games = get_best_or_worst_player_stats(stat='3PM', lowest=True, other_columns=['3PA', '3P%', 'PTS'], n=5)
steph_highest_3pt_games = get_best_or_worst_player_stats(stat='3PM', lowest=False, other_columns=['3PA', '3P%', 'PTS'], n=5)

---

### Best Scorers Against GSW

opp_df = df[df.TEAM != 'GSW'] # Exclude GSW players
print(f"Unique teams: {opp_df.TEAM.nunique()}")
print(f"MIN null values: {opp_df.MIN.isnull().sum()}")

# Filter out players with 0 MIN
opp_df = opp_df[opp_df.MIN.notna()]
print(f"MIN null values: {opp_df.MIN.isnull().sum()}")
opp_df.shape

# Highest Points scored by a player against GSW
opp_df.sort_values(by='PTS', ascending=False)[['NAME', 'TEAM', 'DATE', 'FG%', '3P%', 'FT%', 'PTS']].head(10)

# Highest Opponent Player PPG (4 games or more) OPTION 1
games_played = opp_df.groupby('ID').size().reset_index(name='GAMES')

excl_cols = ['index', 'Unnamed: 0','NAME', 'TEAM', 'OPP', 'STATUS', 'DATE', 'MIN']
incl_cols = [col for col in opp_df.columns if col not in excl_cols]
info_cols = ['ID', 'NAME']

opp_avg = opp_df.groupby(by=['ID'], as_index=False)[incl_cols].mean()

player_info = opp_df[info_cols].drop_duplicates(subset='ID')
new_opp_avg_df = pd.merge(opp_avg, games_played, on='ID')
new_opp_avg_df = pd.merge(new_opp_avg_df, player_info, on='ID')

new_cols = info_cols + [col for col in new_opp_avg_df.columns if col not in info_cols]
new_opp_avg_df = new_opp_avg_df[new_cols]
new_opp_avg_df[new_opp_avg_df['GAMES'] >= 4].sort_values(by=['PTS', 'FG%'], ascending=False).head(10)

# Highest Opponent Player PPG (4 games or more) OPTION 2
gps = opp_df.groupby(by='ID').size().reset_index(name='GAMES')
opp_player_avg_df = df.groupby(by='ID').agg({
    'NAME': 'first', 'FGM': 'mean', 'FGA': 'mean', 'FG%': 'mean', '3PM': 'mean', '3PA': 'mean',
    '3P%': 'mean', 'FTM': 'mean', 'FTA': 'mean', 'FT%': 'mean', 'OREB': 'mean', 'DREB': 'mean', 'TREB': 'mean',
    'AST': 'mean', 'STL': 'mean', 'BLK': 'mean', 'TOV': 'mean', 'PF': 'mean', 'PTS': 'mean', '+/-': 'mean'
}).reset_index()

# Rounding all numeric columns to 2 decimal places
numeric_cols = opp_player_avg_df.select_dtypes(include='number').columns
opp_player_avg_df[numeric_cols] = opp_player_avg_df[numeric_cols].round(2)

opp_player_avg_stats_df = pd.merge(opp_player_avg_df, gps, on='ID')
opp_player_avg_stats_df[opp_player_avg_stats_df['GAMES'] >= 4].sort_values(by=['PTS', 'FG%'], ascending=False)[['NAME', 'FG%', '3P%', 'FT%', 'PTS']].head(10)

opp_base_cols = ['ID', 'NAME', 'TEAM', 'STATUS', 'DATE', 'MIN']
opp_cols_of_interest = ['FGM', 'FGA', '3PM', '3PA', 'FTM', 'FTA',
    'OREB', 'DREB', 'TREB', 'AST', 'STL', 'BLK', 'TOV', 'PF', 'PTS', '+/-']

for col in opp_cols_of_interest:
    print(opp_df[opp_df[col] == opp_df[col].max()][opp_base_cols + [col]], "\n")

---
