==> main/06_gsw_analyze_ppq_data.ipynb <==
## GSW Points Per Quarter Data

import pandas as pd
pd.set_option('display.max_columns', None)


df = pd.read_excel('../data/gsw_box_score_team_stats.xlsx')

gsw_all_games_df = df[df['TEAM'] == 'GSW']

def show_quarters_ppg(data):
    for i in range(1, 5):
        print(f"Q{i}: {round(data[f'Q{i}'].mean(), 2)} PPG")

show_quarters_ppg(gsw_all_games_df) # All GSW games

gsw_all_home_games_df = gsw_all_games_df[gsw_all_games_df['STATUS'] == 'home']
show_quarters_ppg(gsw_all_home_games_df) # All GSW home games

GSW score more points in the 1st quarter when they are at home.

gsw_all_away_games_df = gsw_all_games_df[gsw_all_games_df['STATUS'] == 'away']
show_quarters_ppg(gsw_all_away_games_df) # All GSW away games

GSW score more points in the 4th quarter when they are on the road.

JIMMY_TRADE = '2025-02-08'
gsw_df_pre_jimmy = gsw_all_games_df[gsw_all_games_df['DATE'] < JIMMY_TRADE]
gsw_df_with_jimmy = gsw_all_games_df[gsw_all_games_df['DATE'] >= JIMMY_TRADE]

show_quarters_ppg(gsw_df_pre_jimmy) # GSW games before Jimmy Butler trade

show_quarters_ppg(gsw_df_with_jimmy) # GSW games since Jimmy Butler trade

show_quarters_ppg(gsw_df_with_jimmy[gsw_df_with_jimmy['STATUS'] == 'home']) # GSW home games since Jimmy Butler trade

show_quarters_ppg(gsw_df_with_jimmy[gsw_df_with_jimmy['STATUS'] == 'away']) # GSW away games since Jimmy Butler trade

---

home_wins = gsw_all_home_games_df[gsw_all_home_games_df['W/L']=='W'].shape[0]
home_losses = gsw_all_home_games_df[gsw_all_home_games_df['W/L']=='L'].shape[0]
print(f"GSW's record at home is {home_wins} wins and {home_losses} losses")

away_wins = gsw_all_away_games_df[gsw_all_away_games_df['W/L']=='W'].shape[0]
away_losses = gsw_all_away_games_df[gsw_all_away_games_df['W/L']=='L'].shape[0]
print(f"GSW's record on the road is {away_wins} wins and {away_losses} losses")

---

opp_all_games_df = df[df['TEAM'] != 'GSW']

def get_highest_lowest_opp_quarter_pts(q_num: int, lowest: bool) -> None:
    highest_lowest_str = 'Lowest' if lowest else 'Highest'
    opp_quarter_pts = opp_all_games_df.sort_values(by=f'Q{q_num}', ascending=lowest).head(5)
    print(f'\nQuarter {q_num} {highest_lowest_str} PTS scored by opponent team against GSW')
    print(opp_quarter_pts[['TEAM', 'STATUS', 'DATE', f'Q{q_num}', 'W/L']])

get_highest_lowest_opp_quarter_pts(q_num=1, lowest=False)
get_highest_lowest_opp_quarter_pts(q_num=1, lowest=True)

get_highest_lowest_opp_quarter_pts(q_num=2, lowest=False)
get_highest_lowest_opp_quarter_pts(q_num=2, lowest=True)

get_highest_lowest_opp_quarter_pts(q_num=3, lowest=False)
get_highest_lowest_opp_quarter_pts(q_num=3, lowest=True)

get_highest_lowest_opp_quarter_pts(q_num=4, lowest=False)
get_highest_lowest_opp_quarter_pts(q_num=4, lowest=True)

---

### Show games where GSW were down going into Q4, but won the game

gsw_all_games_df = df[df['TEAM'] == 'GSW']
opp_all_games_df = df[df['TEAM'] != 'GSW']

print(f"GSW df shape {gsw_all_games_df.shape}; expected games: 82")
print(f"OPP df shape {opp_all_games_df.shape}; expected games: 82")

# Merge by date so that each game has both teams' stats in one row
merged_df = pd.merge(gsw_all_games_df, opp_all_games_df, on="DATE", suffixes=("_GSW", "_OPP"))
merged_df.head(2)

# Apply the filtering condition
gsw_down_after_q3_won_game = merged_df[
    (merged_df["Q1_GSW"] + merged_df["Q2_GSW"] + merged_df["Q3_GSW"] <
     merged_df["Q1_OPP"] + merged_df["Q2_OPP"] + merged_df["Q3_OPP"]) &
    (merged_df["W/L_GSW"] == "W")
]

# Display the relevant columns
print("Games where GSW were down going into Q4, but won the game:\n", gsw_down_after_q3_won_game[["TEAM_GSW", "OPP_GSW", "DATE", "W/L_GSW", "STATUS_GSW"]])

### Show games where GSW were up going into Q4, but lost the game

gsw_up_after_q3_lost_game = merged_df[
    (merged_df["Q1_GSW"] + merged_df["Q2_GSW"] + merged_df["Q3_GSW"] >
     merged_df["Q1_OPP"] + merged_df["Q2_OPP"] + merged_df["Q3_OPP"]) &
    (merged_df["W/L_GSW"] == "L")
]
print("Games where GSW had the lead going into Q4, but lost the game:\n", gsw_up_after_q3_lost_game[["TEAM_GSW", "OPP_GSW", "DATE", "W/L_GSW", "STATUS_GSW"]])

### Show games where GSW were up going into Q4, and won the game

gsw_up_after_q3_won_game = merged_df[
    (merged_df["Q1_GSW"] + merged_df["Q2_GSW"] + merged_df["Q3_GSW"] >
     merged_df["Q1_OPP"] + merged_df["Q2_OPP"] + merged_df["Q3_OPP"]) &
    (merged_df["W/L_GSW"] == "W")
]
print(f"Number of games '{gsw_up_after_q3_won_game.shape[0]}'")
print("Games where GSW had the lead going into Q4, and won the game:\n", gsw_up_after_q3_won_game[["TEAM_GSW", "OPP_GSW", "DATE", "W/L_GSW", "STATUS_GSW"]])

---
