==> main/07_playoff_jimmy.ipynb <==
import pandas as pd
pd.set_option('display.max_columns', None)

regular_df = pd.read_excel("../data/jimmy_butler_totals_regular.xlsx")
regular_df.head()

playoffs_df = pd.read_excel("../data/jimmy_butler_totals_playoffs.xlsx")
playoffs_df.head()

playoffs_df.shape

len(playoffs_df)

for col in playoffs_df[['PTS', 'REB', 'AST', 'BLK', 'STL', 'TOV']]:
    print(col)

pl_copy = playoffs_df.copy()

pl_copy['PTS30M'] = pl_copy['PTS'] / (pl_copy['MIN'] / 30)
pl_copy['3PM30M'] = pl_copy['3PM'] / (pl_copy['MIN'] / 30)
pl_copy['FTM30M'] = pl_copy['FTM'] / (pl_copy['MIN'] / 30)
pl_copy['REB30M'] = pl_copy['REB'] / (pl_copy['MIN'] / 30)
pl_copy['AST30M'] = pl_copy['AST'] / (pl_copy['MIN'] / 30)
pl_copy['BLK30M'] = pl_copy['BLK'] / (pl_copy['MIN'] / 30)
pl_copy['STL30M'] = pl_copy['STL'] / (pl_copy['MIN'] / 30)
pl_copy['TOV30M'] = pl_copy['TOV'] / (pl_copy['MIN'] / 30)

pl_copy.head()
